+++ gui/shapes/OtCurvePath.h
@@ -119,9 +119,9 @@ private:
 	// find the segment at u [0, 1]
 	OtCurve findSegment(float u, float* fraction) {
 		float target = currentLength * u;
-		int low = 0;
-		int high = lengths.size() - 1;
-		int i;
+		size_t low = 0;
+		size_t high = lengths.size() - 1;
+		size_t i;
 
 		while (low <= high) {
 			i = (low + high) / 2;
